feat : add table with API support (#1)
Create the HTML table which will be the main attraction of the application. In this first form, the table is plugged into the CheapShark API (https://apidocs.cheapshark.com). Currently we pull the Deals data from CheapShark and we present them to the user. The table can be sorted by Store Name, Price, Title, Release date and Reviews by clicking on the respective item on the table's head. At the end of the table, the user will find a Load More button, which when clicked, will make a new GET request and fetch 60 more games.