feat : make the table responsive (#2)
Made the table to be responsive up to 320px of screen width. When reaching a screen width of 980px the table rows stack onto each other giving the illusion of being a card item ( but we still retain the table structure ). In order for the items to be more presentable, on the mobile view the border for each row is removed and we add a horizontal line ( hr element ) so we can space each item and create an easy to read / see experience.