feat : update table styles (#3)
Updated the styling of the table. Before there was too much information on a single row. Now each row contains the basic information about a game. Planning to add a feature which will allow the user to hover on a single item and more details about the game will appear.